+++ packages/emails/tests/smtp.test.ts
@@ -1,6 +1,6 @@
 /* eslint-disable import/no-extraneous-dependencies */
-import url from 'url';
-import fs from 'fs';
+import url from 'node:url';
+import fs from 'node:fs';
 import { test, expect } from 'vitest';
 import email, { smtp as smtpSend } from '../src/index';