+++ packages/feeds/src/firebase.ts
@@ -1,17 +1,15 @@
 /* eslint-disable import/prefer-default-export, import/first */
-import { initializeApp } from 'firebase-admin/app';
-
-initializeApp();
-
 import { onRequest } from 'firebase-functions/v2/https';
 import config from '@cloudcommerce/firebase/lib/config';
 import serveFeeds from './firebase/serve-feeds';
 
-const { ssrFunctionOptions } = config.get();
+const {
+  httpsFunctionOptions: { region },
+} = config.get();
 
 export const feeds = onRequest({
-  concurrency: 80,
-  ...ssrFunctionOptions,
+  concurrency: 300,
+  region,
   memory: '512MiB',
   timeoutSeconds: 120,
 }, serveFeeds);