+++ packages/storefront/src/lib/ssr-context.ts
@@ -24,6 +24,12 @@ type StorefrontConfig = {
 const emitter = new EventEmitter();
 const getConfig: () => StorefrontConfig = _getConfig;
 
+if (!globalThis.$apiMergeConfig) {
+  globalThis.$apiMergeConfig = {
+    isNoAuth: true,
+    canCache: true,
+  };
+}
 if (!globalThis.$apiPrefetchEndpoints) {
   globalThis.$apiPrefetchEndpoints = ['categories'];
 }
@@ -71,13 +77,9 @@ const loadPageContext = async (Astro: Readonly<AstroGlobal>, {
     brands?: BrandsList,
     [k: string]: Record<string, any> | undefined,
   } = {};
-  const apiOptions = {
-    fetch,
-    isNoAuth: true,
-  };
   const apiFetchings = [
     null, // fetch by slug
-    ...apiPrefetchEndpoints.map((endpoint) => api.get(endpoint, apiOptions)),
+    ...apiPrefetchEndpoints.map((endpoint) => api.get(endpoint)),
   ];
   if (isHomepage) {
     cmsContent = await config.getContent('home');
@@ -86,8 +88,13 @@ const loadPageContext = async (Astro: Readonly<AstroGlobal>, {
     if (slug) {
       if (contentCollection) {
         cmsContent = await config.getContent(`${contentCollection}/${slug}`);
+      } else if (slug.startsWith('api/')) {
+        const err: any = new Error('/api/* routes not implemented on SSR directly');
+        Astro.response.status = 501;
+        err.responseHTML = `<head></head><body>${err.message}</body>`;
+        throw err;
       } else {
-        apiFetchings[0] = api.get(`slugs/${slug}`, apiOptions);
+        apiFetchings[0] = api.get(`slugs/${slug}`);
       }
     }
   }